requirements.txt
python-telegram-bot==20.7
psycopg==3.2.3
python-dotenv==1.0.0 